done faire quelque test + trouver scenario

+++ TransconnectProject/Model/ProduitModel/Produit.cs
@@ -21,7 +21,7 @@ namespace TransconnectProject.Model.ProduitModel
 		public static Produit createProduit()
 		{
 			string priceKg ="";
-            int kg;
+            double kg;
             Console.WriteLine("\n- Fonctionnalite de creation de Produit -\n");
             Console.Write("veuillez saisir le nom du Produit: ");
             string nom = Console.ReadLine();
@@ -31,7 +31,7 @@ namespace TransconnectProject.Model.ProduitModel
                 Console.Write("Veuillez saisir le Prix au kilo: ");
                 priceKg = Console.ReadLine();
 				Console.Clear();
-            } while (!int.TryParse(priceKg,out kg));
+            } while (!double.TryParse(priceKg,out kg));
             
             Produit p = new Produit(nom,kg);
 			if(p!=null)
